+++ R/assertions.R
@@ -98,11 +98,7 @@ assert_learners = function(learners, task = NULL, properties = character(0L), cl
 #' @export
 #' @param measure :: [Measure].
 #' @rdname mlr_assertions
-assert_measure = function(measure, task_type = NULL, task = NULL, learner = NULL, clone = FALSE) {
-  task_type = task_type %??% task$task_type %??% learner$task_type
-  if (is.null(measure)) {
-    measure = mlr_reflections$default_measures[[task_type]]
-  }
+assert_measure = function(measure, task = NULL, learner = NULL, clone = FALSE) {
   measure = cast_from_dict(measure, "Measure", mlr_measures, clone, FALSE)[[1L]]
 
   if (!is.null(task)) {
@@ -131,18 +127,14 @@ assert_measure = function(measure, task_type = NULL, task = NULL, learner = NULL
 #' @export
 #' @param measures :: list of [Measure].
 #' @rdname mlr_assertions
-assert_measures = function(measures, task_type = NULL, task = NULL, learner = NULL, clone = FALSE) {
-  task_type = task_type %??% task$task_type %??% learner$task_type
-  if (is.null(measures)) {
-    measures = list(mlr_reflections$default_measures[[task_type]])
-  } else {
-    measures = cast_from_dict(measures, "Measure", mlr_measures, clone, TRUE)
-  }
+assert_measures = function(measures, task = NULL, learner = NULL, min_len = 0L, clone = FALSE) {
+  measures = cast_from_dict(measures, "Measure", mlr_measures, clone, TRUE)
 
-  if (length(measures) == 0L) {
-    stopf("You must provide at least one measure")
+  if (min_len > length(measures)) {
+    if (min_len == 1L)
+      stopf("You must provide at least one measures")
+    stopf("You must provide at least %i measures", min_len)
   }
-
   lapply(measures, assert_measure, task = task, learner = learner)
 }