+++ tests/testthat/test_DataBackendRbind.R
@@ -55,7 +55,7 @@ test_that("Backends with mixed data_formats", {
   expect_set_equal(task$backend$data_formats, "data.table")
 
   rr = resample(task, "regr.rpart", "holdout")
-  expect_number(rr$aggregate())
+  expect_number(rr$aggregate("regr.mse"))
 })
 
 test_that("Backends with same rows", {